requirements.txt
beautifulsoup4==4.9.1
Django==2.2.15
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-localflavor==3.0.1
django-phone-field==1.8.1
importlib-metadata==1.7.0
python-stdnum==1.13
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
